requirements.txt
django==2.2.10
djangorestframework==3.9.2
# patched to add get-or-create semantics for nested objects
https://github.com/claytondaley/drf-writable-nested/zipball/b49aefae8a0bc7ce763e007ab3d474d11cfd4b07/
